FEAT: Working actuator class. with thruster and fins. renamed some variables
